requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
colorama==0.4.6
fastapi==0.110.0
firebird-base==1.7.2
firebird-driver==1.10.1
firebirdsql==1.2.5
h11==0.14.0
idna==3.6
protobuf==5.26.0
pydantic==2.6.3
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.28.0
websockets==12.0
